+++ R/pmsampsize.R
@@ -1,9 +1,10 @@
 #' pmsampsize
-#' - Calculates the minimum sample size required for developing a multivariable prediction model
+#' - Sample Size for Development of a Prediction Model
 #'
-#' \code{pmsampsize} computes the minimum sample size required for the development of a new
-#' multivariable prediction model using the criteria proposed by Riley \emph{et al}. 2018. \code{pmsampsize}
-#' can be used to calculate the minimum sample size for the development of models with
+#' @description \code{pmsampsize} computes the minimum sample size required for the development of a new
+#' multivariable prediction model using the criteria proposed by Riley \emph{et al}. 2018.
+#'
+#' @details \code{pmsampsize} can be used to calculate the minimum sample size for the development of models with
 #' continuous, binary or survival (time-to-event) outcomes. Riley \emph{et al}. lay out a series of
 #' criteria the sample size should meet. These aim to minimise the overfitting and to ensure
 #' precise estimation of key parameters in the prediction model. \cr \cr
@@ -22,16 +23,14 @@
 #' iii) precise estimation (within +/- 0.05) of the average outcome risk in the
 #' population for a key timepoint of interest for prediction.
 #'
-#' NB: When specifying a binary outcome prediction model with 12 or fewer predictor
-#' parameters, an alternative approach by van Smeden et al. may be considered as
-#' presented here: https://mvansmeden.shinyapps.io/BeyondEPV/
 #'
-#' With thanks to Gary Collins, Glen Martin & Kym Snell for helpful input & feedback
+#' With thanks to Richard D. Riley, Emma C Martin, Gary Collins, Glen Martin & Kym Snell for helpful input & feedback
 #'
+#' @return A list including a matrix of calculated sample size requirements for each criteria defined
+#' under \emph{'Details'}, and a series of vectors of parameters used in the calculations as well as the
+#' final recommended minimum sample size and number of events required for model development.
 #'
-#' @author Joie Ensor (Keele University, j.ensor@keele.ac.uk),
-#' @author Emma C. Martin (Phastar),
-#' @author Richard D. Riley (Keele University)
+#' @author Joie Ensor (University of Birmingham, j.ensor@bham.ac.uk),
 #'
 #' @param type specifies the type of analysis for which sample size is being calculated
 #'      \itemize{
@@ -39,22 +38,34 @@
 #'          \item \code{"b"} specifies sample size calculation for a prediction model with a binary outcome
 #'          \item \code{"s"} specifies sample size calculation for a prediction model with a survival (time-to-event) outcome
 #'      }
-#' @param rsquared specifies the expected value of the (Cox-Snell) R-squared of the new model,
+#'
+#' @param rsquared for \code{type="c"} this specifies the expected value of the R-squared of the new model,
 #' where R-squared is the percentage of variation in outcome values explained by the model.
-#' For example, the user may input the value of the (Cox-Snell) R-squared reported for a
+#' For example, the user may input the value of the R-squared reported for a
 #' previous prediction model study in the same field.  If taking a value from a previous
 #' prediction model development study, users should input the model's adjusted R-squared
 #' value, not the apparent R-squared value, as the latter is optimistic (biased).  However,
 #' if taking the R-squared value from an external validation of a previous model, the
 #' apparent R-squared can be used (as the validation data was not used for development, and
-#' so R-squared apparent is then unbiased).  Note that for binary and survival outcome
-#' models, the Cox-Snell R-squared value is required; this is the generalised version of
-#' the well-known R-squared for continuous outcomes, based on the likelihood.  The papers
-#' by Riley et al. (see references) outline how to obtain the Cox-Snell R-squared value
-#' from published studies if they are not reported, using other information (such as the
-#' C-statistic [see cstatistic() option below] or Nagelkerke's R-squared).  Users should
-#' be conservative with their chosen R-squared value; for example, by taking the R-squared
+#' so R-squared apparent is then unbiased).
+#' Users should be conservative with their chosen R-squared value; for example, by taking the R-squared
 #' value from a previous model, even if they hope their new model will improve performance.
+#'
+#' @param csrsquared for \code{type="b" or type="s"} this specifies the expected value of
+#' the Cox-Snell R-squared of the new model. The Cox-Snell R-squared is the generalised
+#' version of the well-known R-squared for continuous outcomes, based on the likelihood.
+#' Please read the description of \code{rsquared} for additional details about specifying the
+#' expected R-squared performance.
+#' The papers by Riley et al. (see references) outline how to obtain the Cox-Snell R-squared value
+#' from published studies if they are not reported, using other information (such as the
+#' C-statistic [see \code{cstatistic()} option below]).
+#'
+#' @param nagrsquared for \code{type="b" or type="s"} this specifies the expected value of
+#' the Nagelkerke's R-squared of the new model, which is the Cox-Snell R-squared scaled to
+#' lie in the [0,1] range. It is interpretable in the same way as the standard R-squared, i.e.
+#' the percentage of variation in outcome values explained by the model.
+#' Please read the description of \code{rsquared} for additional details about specifying the
+#' expected R-squared performance
 #'
 #' @param parameters specifies the number of candidate predictor parameters for potential
 #' inclusion in the new prediction model.  Note that this may be larger than the number of
@@ -68,45 +79,45 @@
 #' the model would need to be shrunk by 10\% to adjust for overfitting. See references
 #' below for further information.
 #'
-#' @param prevalence (binary outcome option) specifies the overall outcome proportion
+#' @param prevalence (\code{type="b"} option) specifies the overall outcome proportion
 #' (for a prognostic model) or
 #' overall prevalence (for a diagnostic model) expected within the model development
 #' dataset. This should be derived based on previous studies in the same population.
 #'
-#' @param cstatistic (binary outcome option) specifies the C-statistic reported in an
+#' @param cstatistic (\code{type="b"} option) specifies the C-statistic reported in an
 #' existing prediction model study to be used in conjunction with the expected
 #' prevalence to approximate the Cox-Snell R-squared using the approach of Riley et al. 2020.
 #' Ideally, this should be an optimism-adjusted C-statistic. The approximate Cox-Snell R-squared
-#' value is used as described above for the rsquared() option, and so is treated as a baseline
+#' value is used as described above for the \code{csrsquared()} option, and so is treated as a baseline
 #' for the expected performance of the new model.
 #'
-#' @param seed (binary outcome option) specifies the initial value of the random-number
+#' @param seed (\code{type="b"} option) specifies the initial value of the random-number
 #' seed used by the random-number functions when simulating data to approximate the
 #' Cox-Snell R-squared based on reported C-statistic and expect prevalence as described
 #' by Riley et al. 2020
 #'
-#' @param rate (survival outcome option) specifies the overall event rate in the population of interest,
+#' @param rate (\code{type="s"} option) specifies the overall event rate in the population of interest,
 #' for example as obtained from a previous study, for the survival outcome of interest. NB: rate must
 #' be given in time units used for meanfup and timepoint options.
 #'
-#' @param timepoint (survival outcome option) specifies the timepoint of interest for prediction.
+#' @param timepoint (\code{type="s"} option) specifies the timepoint of interest for prediction.
 #' NB: time units must be the same as given for meanfup option (e.g. years, months).
 #'
-#' @param meanfup (survival outcome option) specifies the average (mean) follow-up time
+#' @param meanfup (\code{type="s"} option) specifies the average (mean) follow-up time
 #' anticipated for individuals in the model development dataset,
 #' for example as taken from a previous study in the population of interest.
 #' NB: time units must be the same as given for timepoint option.
 #'
-#' @param  intercept (continuous outcome options) specifies the average outcome value in the population of
+#' @param  intercept (\code{type="c"} options) specifies the average outcome value in the population of
 #' interest e.g. the average blood pressure, or average pain score.
 #' This could be based on a previous study, or on clinical knowledge.
 #'
-#' @param sd (continuous outcome options) specifies the standard deviation (SD) of
+#' @param sd (\code{type="c"} options) specifies the standard deviation (SD) of
 #' outcome values in the population e.g.
 #' the SD for blood pressure in patients with all other predictors set to the average.
 #' This could again be based on a previous study, or on clinical knowledge.
 #'
-#' @param mmoe (continuous outcome options) multiplicative margin of error (MMOE)
+#' @param mmoe (\code{type="c"} options) multiplicative margin of error (MMOE)
 #' acceptable for calculation of the
 #' intercept. The default is a MMOE of 10\%. Confidence interval for the intercept will be
 #' displayed in the output for reference. See references below for further information.
@@ -126,7 +137,7 @@
 #' # Cox-Snell R-squared of an existing prediction model) for the new model's
 #' # R-squared value is 0.288
 #'
-#' pmsampsize(type = "b", rsquared = 0.288, parameters = 24, prevalence = 0.174)
+#' pmsampsize(type = "b", csrsquared = 0.288, parameters = 24, prevalence = 0.174)
 #'
 #' # Now lets assume we could not obtain a Cox-Snell R-squared estimate from an existing
 #' # prediction model, but instead had a C-statistic (0.89) reported for the existing prediction
@@ -145,7 +156,7 @@
 #' # timepoint of interest for prediction using the newly developed model of 2
 #' # years
 #'
-#' pmsampsize(type = "s", rsquared = 0.051, parameters = 30, rate = 0.065,
+#' pmsampsize(type = "s", csrsquared = 0.051, parameters = 30, rate = 0.065,
 #'            timepoint = 2, meanfup = 2.07)
 #'
 #' ## Continuous outcomes (Linear prediction models)
@@ -170,10 +181,6 @@
 #' Part II binary and time-to-event outcomes.
 #' \emph{Statistics in Medicine}. 2018 (in-press). doi: 10.1002/sim.7992
 #'
-#' @references van Smeden M, Moons KG, de Groot JA, et al. Sample size for binary logistic
-#' prediction models: Beyond events per variable criteria.
-#' \emph{Stat Methods Med Res}. 2019;28(8):2455-74
-#'
 #' @references Riley, RD, Van Calster, B, Collins, GS. A note on estimating the Cox-Snell R2
 #' from a reported C statistic (AUROC) to inform sample size calculations for developing a
 #' prediction model with a binary outcome. \emph{Statistics in Medicine}. 2020
@@ -181,6 +188,8 @@
 
 #' @export
 pmsampsize <- function(type,
+                       nagrsquared = NA,
+                       csrsquared = NA,
                        rsquared = NA,
                        parameters,
                        shrinkage = 0.9,
@@ -195,7 +204,7 @@ pmsampsize <- function(type,
                        mmoe=1.1) {
 
   # error checking
-  pmsampsize_errorcheck(type=type,rsquared=rsquared,parameters=parameters,shrinkage=shrinkage,cstatistic=cstatistic,
+  pmsampsize_errorcheck(type=type,nagrsquared=nagrsquared,csrsquared=csrsquared,rsquared=rsquared,parameters=parameters,shrinkage=shrinkage,cstatistic=cstatistic,
             prevalence=prevalence,rate=rate,timepoint=timepoint,meanfup=meanfup,intercept=intercept,sd=sd,mmoe=mmoe)
 
   # choose function based on analysis type
@@ -203,10 +212,31 @@ pmsampsize <- function(type,
                                    sd=sd,shrinkage=shrinkage,mmoe=mmoe)
 
   if (type == "b") {
+    rsquared <- NA
+    if (is.na(nagrsquared) == F) {
+
+      if (is.na(csrsquared) == F) {
+        stop("Only one of csrsquared() or nagrsquared() can be specified")
+      }
+
+       E <- parameters*prevalence
+			 lnLnull <- (E*(log(E/parameters)))+((parameters-E)*(log(1-(E/parameters))))
+       max_r2a <- (1- exp((2*lnLnull)/parameters))
+			 rsquared <- round(nagrsquared*max_r2a,digits=3)
+
+    }
+
+    if (is.na(csrsquared) == F) {
+      if (is.na(nagrsquared) == F) {
+        stop("Only one of csrsquared() or nagrsquared() can be specified")
+      }
+      rsquared <- csrsquared
+    }
+
     if (is.na(cstatistic) == F) {
 
       if (is.na(rsquared) == F) {
-        stop("Only one of rsquared() or cstatistic() can be specified")
+        stop("Only one of csrsquared() or nagrsquared() or cstatistic() can be specified")
       }
 
       approx_rsq <- cstat2rsq(cstatistic=cstatistic, prevalence=prevalence, seed=seed)
@@ -214,15 +244,43 @@ pmsampsize <- function(type,
     }
     else {
       if (is.na(rsquared)) {
-        stop("One of rsquared() or cstatistic() must be specified")
+        stop("One of csrsquared() or nagrsquared() or cstatistic() must be specified")
       }
     }
     out <- pmsampsize_bin(rsquared=rsquared,parameters=parameters,prevalence=prevalence,
                                          shrinkage=shrinkage,cstatistic=cstatistic)
   }
 
-  if (type == "s") out <- pmsampsize_surv(rsquared=rsquared,parameters=parameters,rate=rate,
+  if (type == "s") {
+    rsquared <- NA
+    if (is.na(nagrsquared) == F) {
+
+      if (is.na(csrsquared) == F) {
+        stop("Only one of csrsquared() or nagrsquared() can be specified")
+      }
+
+       events <- parameters*rate*meanfup
+       lnLnull <- (events*(log(events/parameters)))-events
+		 max_r2a <- (1- exp((2*lnLnull)/parameters))
+      rsquared <- round(nagrsquared*max_r2a,digits=3)
+
+    }
+
+    if (is.na(csrsquared) == F) {
+      if (is.na(nagrsquared) == F) {
+        stop("Only one of csrsquared() or nagrsquared() can be specified")
+      }
+      rsquared <- csrsquared
+    }
+    else {
+      if (is.na(rsquared)) {
+        stop("One of csrsquared() or nagrsquared() must be specified")
+      }
+    }
+
+    out <- pmsampsize_surv(rsquared=rsquared,parameters=parameters,rate=rate,
                                           timepoint=timepoint,meanfup=meanfup,shrinkage=shrinkage,mmoe=mmoe)
+  }
 
 
   est <- out